rag app/requirements.txt
langchain
langchain-community
langchain-google-genai
google-generativeai
pypdf
qdrant-client
fastapi
uvicorn
python-dotenv
